+++ modules/voice/mock_action_dispatcher.h
@@ -1,0 +1,14 @@
+#pragma once
+
+#include "modules/voice/action_dispatcher.h"
+
+namespace cockpit {
+namespace voice {
+
+class MockActionDispatcher final : public ActionDispatcher {
+ public:
+  ActionExecutionResult Execute(VoiceAction action) override;
+};
+
+}  // namespace voice
+}  // namespace cockpit